+++ src/read.rs
@@ -3,7 +3,7 @@ use std::fs::File;
 use std::io::BufRead;
 use std::io::BufReader;
 
-pub fn read(side: u8, fic: &str, debug: bool) -> Grid {
+pub fn read(side: u16, fic: &str, debug: bool) -> Grid {
     let mut g = Grid::new(side);
     g.set_debug(debug);
     let input = File::open(&fic);
@@ -25,7 +25,7 @@ pub fn read(side: u8, fic: &str, debug: bool) -> Grid {
     g
 }
 
-pub fn from_vec(side: u8, data: Vec<String>, debug: bool) -> Grid {
+pub fn from_vec(side: u16, data: Vec<String>, debug: bool) -> Grid {
     let mut g = Grid::new(side);
     g.set_debug(debug);
     let mut line_number = 1;
@@ -36,7 +36,7 @@ pub fn from_vec(side: u8, data: Vec<String>, debug: bool) -> Grid {
     g
 }
 
-pub fn from_vecvec(side: u8, data: &[Vec<u8>], debug: bool) -> Grid {
+pub fn from_vecvec(side: u16, data: &[Vec<u16>], debug: bool) -> Grid {
     let mut g = Grid::new(side);
     g.set_debug(debug);
     g.compute_vecvec(data);
